requirements.txt
Django==2.2.5
djangorestframework==3.10.3
Pillow==6.1.0
psycopg2==2.8.3
pytest-django==3.5.1